requirements.txt
faiss-cpu
numpy
sentence-transformers
python-dotenv
pymongo
typer
rich
sounddevice
soundfile
SpeechRecognition
google-generativeai
jinja2
cryptography
pytest
mongomock 